Moving all the test/test_helper files into db/pg as they should also be used as server helpers, also changing the file paths for their tests
