+++ db/pg/updateCollection/handleDeleteWords.js
