+++ db/pg/updateCollection/handleNewWords.js
