+++ db/pg/updateCollection/handleUpdateWords.js
