+++ test/db/createCollection.test.js
@@ -1,13 +1,12 @@
 const tape = require('tape');
 const assert = require('assert');
-
-const redisCli = require('./configureRedis.js');
-const pg = require('./configurePool.js');
+const redisCli = require('../test_helpers/configureRedis.js');
+const pg = require('../test_helpers/configurePool.js');
 const pool = new pg.pool(pg.config);
 
 const flushDb = require('../helpers/flushDb.js')(pool, redisCli);
 const registerUser = require('../helpers/registerUser.js')(pool, redisCli);
-const createCollection = require('../helpers/createCollection.js')(pool);
+const createCollection = require('../../db/pg/createCollection.js')(pool);
 
 function getAllCollections() {
  return new Promise((resolve) => {
