+++ test/db/index.test.js
@@ -1,2 +1,11 @@
 require('./setDefinedString.test.js');
 require('./setDefinedValues.test.js');
+
+require('./createCollection.test.js');
+require('./getCollections.test.js');
+require('./updateCollection.test.js');
+require('./updateCollection/handleCollectionUpdates.test.js');
+require('./updateCollection/handleNewWords.test.js');
+require('./updateCollection/handleUpdateWords.test.js');
+require('./updateCollection/handleDeleteWords.test.js');
+require('./getWords.test.js');
