+++ test/server/api/collection.test.js
@@ -7,8 +7,8 @@ const pool = server.app.pool;
 
 const flushDb = require('../../helpers/flushDb.js')(pool, redisCli);
 const authenticate = require('../../helpers/authenticate.js')(pool, redisCli);
-const getCollections = require('../../helpers/getCollections.js')(pool);
-const createCollection = require('../../helpers/createCollection.js')(pool);
+const getCollections = require('../../../db/pg/getCollections.js')(pool);
+const createCollection = require('../../../db/pg/createCollection.js')(pool);
 
 tape('GET :: /api/collection', (t) => {
   const options = {
