+++ test/test_helpers/index.test.js
@@ -4,12 +4,3 @@ require('./loginUserWRedis.test.js');
 require('./registerUser.test.js');
 require('./flushDb.test.js');
 require('./authenticate.test.js');
-
-require('./createCollection.test.js');
-require('./getCollections.test.js');
-require('./updateCollection.test.js');
-require('./updateCollection/handleCollectionUpdates.test.js');
-require('./updateCollection/handleNewWords.test.js');
-require('./updateCollection/handleUpdateWords.test.js');
-require('./updateCollection/handleDeleteWords.test.js');
-require('./getWords.test.js');
